// frontend/src/app/models/korisnik.ts
export class Korisnik {
  korisnicko_ime: string = ''
  lozinka: string = ''
  ime: string = ''
  prezime: string = ''
  pol: string = ''
  adresa: string = ''
  telefon: string = ''
  email: string = ''
  profilna_slika: string = ''
  broj_kartice: string = ''
  tip: string = ''
  status: string = ''
  firma: string = ''
  brojZakPoslova: number = 0
}